Created a token list class
i built a class that will accept ArrayList<String> arguments and will identify the frequency and print out to the console the number of duplicates. Its a little rough now so i can tweak it later so we can sort it by most and least occurring tokens. But it works.