This is the source code from the lecture on Thursday 10/11
its not complete, its as much as i could get written down. It has the structure for Node and NodeEntry as well as some sample unit tests in the AppTest file. The FriendlyMap class header also has the proper extension and implementation signature for the assignment, the one he said to use.